Split menu mutations
