+++ saleor/graphql/menu/mutations/__init__.py
@@ -1,0 +1,19 @@
+from .assign_navigation import AssignNavigation
+from .menu_create import MenuCreate
+from .menu_delete import MenuDelete
+from .menu_item_create import MenuItemCreate
+from .menu_item_delete import MenuItemDelete
+from .menu_item_move import MenuItemMove
+from .menu_item_update import MenuItemUpdate
+from .menu_update import MenuUpdate
+
+__all__ = [
+    "AssignNavigation",
+    "MenuCreate",
+    "MenuDelete",
+    "MenuItemCreate",
+    "MenuItemDelete",
+    "MenuItemMove",
+    "MenuItemUpdate",
+    "MenuUpdate",
+]
